Part 12 complete

+++ src/entities/Entity.java
@@ -18,20 +18,14 @@ public abstract class Entity {
     }
 
     protected void drawHitbox(Graphics g) {
-        // For debugging the hitbox
         g.setColor(Color.pink);
         g.drawRect((int) hitbox.x, (int) hitbox.y, (int) hitbox.width, (int) hitbox.height);
     }
 
-    protected void initHitbox(float x, float y, float width, float height) {
+    protected void initHitbox(float x, float y, int width, int height) {
         hitbox = new Rectangle2D.Float(x, y, width, height);
 
     }
-
-//    protected void updateHitbox() {
-//        hitbox.x = (int) x;
-//        hitbox.y = (int) y;
-//    }
 
     public Rectangle2D.Float getHitbox() {
         return hitbox;
